接收CTP地锁数据,存储停车入场数据 (#37)

+++ shared-trad-data-serv/shared-trad-data-impl/src/main/java/cn/suparking/data/service/ParkingEventService.java
@@ -16,5 +16,5 @@ public interface ParkingEventService {
      * @param parkingEventDTO {@link ParkingEventDTO}
      * @return int
      */
-    Integer createOrUpdate(ParkingEventDTO parkingEventDTO);
+    Long createOrUpdate(ParkingEventDTO parkingEventDTO);
 }
